Pages should be text files,
trying opening it up as utf-8 file (i.e. if it's ascii text, then utf-8 covers ascii)

If it fails, then try opening it as latin-1. If it fails after that, then just fail as normal.

I may be able to revert all the unicode conversions sprinkled all over the place?

Of course, this particular bit of transgression fails one of the tests (because the open sees a \r\n, whilst regular file.open only saw an \n.